Implements metrics.
This commit implements the requested metrics, which involve calculating the average time it takes tasks to get done,
both in general and by priority. The new class with the implemented methods is /utils/ToDoMetrics.java, and the
files modified in this commit are ToDo.java, ToDoServiesImpl.java and ToDoController.java.